Feature/user experience

+++ src/components/molecules/TransactionList.tsx
@@ -1,5 +1,5 @@
 import {Animated, StyleSheet, TouchableOpacity, View} from 'react-native';
-import React, {useRef} from 'react';
+import React, {useState} from 'react';
 import useThemeColors, {Colors} from '../../hooks/useThemeColors';
 import Icon from '../atoms/Icons';
 import moment from 'moment';
@@ -13,6 +13,8 @@ import Expense from '../../schemas/ExpenseSchema';
 import {Dispatch} from 'redux';
 import PrimaryText from '../atoms/PrimaryText';
 import {getEverydayExpenseRequest} from '../../redux/slice/everydayExpenseDataSlice';
+import UndoModal from '../atoms/UndoModal';
+import {formatCurrency} from '../../utils/numberUtils';
 
 interface TransactionListProps {
   currencySymbol: string;
@@ -29,15 +31,27 @@ interface TransactionItemProps {
   targetDate?: string;
 }
 
+const truncateString = (str, maxLength) => {
+  if (str.length > maxLength) {
+    return str.substring(0, maxLength) + '..';
+  } else {
+    return str;
+  }
+};
+
 const TransactionItem: React.FC<TransactionItemProps> = ({
   currencySymbol,
-  expense,
+  expense: initialExpense,
   colors,
   dispatch,
   label,
   targetDate,
 }) => {
-  const slideAnim = useRef(new Animated.Value(1)).current;
+  const [expenses, setExpenses] = useState<Array<Expense>>(initialExpense);
+  const [deletedItem, setDeletedItem] = useState<Expense | null>(null);
+  const [deletionTimeoutId, setDeletionTimeoutId] = useState<number | null>(
+    null,
+  );
 
   const handleEdit = (
     expenseId: string,
@@ -47,32 +61,46 @@ const TransactionItem: React.FC<TransactionItemProps> = ({
     expenseDate: string,
     expenseAmount: number,
   ) => {
-    Animated.timing(slideAnim, {
-      toValue: 200,
-      duration: 250,
-      useNativeDriver: true,
-    }).start(() => {
-      navigate('UpdateTransactionScreen', {
-        expenseId,
-        expenseTitle,
-        expenseDescription,
-        category,
-        expenseDate,
-        expenseAmount,
-      });
+    navigate('UpdateTransactionScreen', {
+      expenseId,
+      expenseTitle,
+      expenseDescription,
+      category,
+      expenseDate,
+      expenseAmount,
     });
   };
 
   const handleDelete = (expenseId: string) => {
-    Animated.timing(slideAnim, {
-      toValue: -200,
-      duration: 250,
-      useNativeDriver: true,
-    }).start(() => {
-      deleteExpenseById(Realm.BSON.ObjectID.createFromHexString(expenseId));
-      dispatch(getExpenseRequest());
-      dispatch(getEverydayExpenseRequest(targetDate));
-    });
+    const deletedExpense = expenses.find(
+      expense => String(expense._id) === expenseId,
+    );
+    setExpenses(prevExpenses =>
+      prevExpenses.filter(expense => String(expense._id) !== expenseId),
+    );
+    setDeletedItem(deletedExpense);
+
+    const deletionTimeout = setTimeout(() => {
+      if (!deletedItem) {
+        deleteExpenseById(Realm.BSON.ObjectID.createFromHexString(expenseId));
+        dispatch(getExpenseRequest());
+        dispatch(getEverydayExpenseRequest(targetDate));
+      }
+    }, 3000);
+
+    setTimeout(() => {
+      setDeletedItem(null);
+    }, 3000);
+
+    setDeletionTimeoutId(deletionTimeout);
+  };
+
+  const handleUndo = () => {
+    if (deletedItem) {
+      clearTimeout(deletionTimeoutId);
+      setExpenses(prevExpenses => [...prevExpenses, deletedItem]);
+      setDeletedItem(null);
+    }
   };
 
   const renderLeftActions = (expense: Expense) => {
@@ -93,7 +121,7 @@ const TransactionItem: React.FC<TransactionItemProps> = ({
           style={[
             styles.swipeView,
             {
-              backgroundColor: colors.cardBackground,
+              backgroundColor: colors.lightAccent,
               borderTopLeftRadius: 10,
               borderBottomLeftRadius: 10,
             },
@@ -108,7 +136,7 @@ const TransactionItem: React.FC<TransactionItemProps> = ({
         <View
           style={[
             styles.stretchView,
-            {backgroundColor: colors.cardBackground, marginRight: -250},
+            {backgroundColor: colors.lightAccent, marginRight: -250},
           ]}
         />
       </TouchableOpacity>
@@ -123,14 +151,14 @@ const TransactionItem: React.FC<TransactionItemProps> = ({
         <View
           style={[
             styles.stretchView,
-            {backgroundColor: colors.cardBackground, marginLeft: -250},
+            {backgroundColor: colors.lightAccent, marginLeft: -250},
           ]}
         />
         <View
           style={[
             styles.swipeView,
             {
-              backgroundColor: colors.cardBackground,
+              backgroundColor: colors.lightAccent,
               borderTopRightRadius: 10,
               borderBottomRightRadius: 10,
             },
@@ -149,7 +177,7 @@ const TransactionItem: React.FC<TransactionItemProps> = ({
   return (
     <View>
       <PrimaryText style={{fontSize: 12, marginBottom: 5}}>{label}</PrimaryText>
-      {expense.map((expense: Expense) => (
+      {expenses.map((expense: Expense) => (
         <GestureHandlerRootView key={String(expense._id)}>
           <Swipeable
             renderLeftActions={() => renderLeftActions(expense)}
@@ -160,14 +188,13 @@ const TransactionItem: React.FC<TransactionItemProps> = ({
                 styles.transactionContainer,
                 {
                   backgroundColor: colors.containerColor,
-                  // transform: [{translateX: slideAnim}],
                 },
               ]}>
               <View style={styles.iconNameContainer}>
                 <View
                   style={[
                     styles.iconContainer,
-                    {backgroundColor: colors.primaryText},
+                    {backgroundColor: colors.iconContainer},
                   ]}>
                   <Icon
                     name={expense.category.icon ?? 'selection-ellipse'}
@@ -177,7 +204,7 @@ const TransactionItem: React.FC<TransactionItemProps> = ({
                   />
                 </View>
                 <View>
-                  <PrimaryText>{expense.title}</PrimaryText>
+                  <PrimaryText>{truncateString(expense.title, 15)}</PrimaryText>
                   <View style={styles.descriptionContainer}>
                     <PrimaryText
                       style={{
@@ -185,16 +212,20 @@ const TransactionItem: React.FC<TransactionItemProps> = ({
                         fontSize: 10,
                         marginRight: 5,
                       }}>
-                      {expense.category.name} .
+                      {truncateString(expense.category.name, 7)} ◦
                     </PrimaryText>
-                    <PrimaryText
-                      style={{
-                        color: colors.primaryText,
-                        fontSize: 10,
-                        marginRight: 5,
-                      }}>
-                      {expense.description} .
-                    </PrimaryText>
+
+                    {expense.description !== '' ? (
+                      <PrimaryText
+                        style={{
+                          color: colors.primaryText,
+                          fontSize: 10,
+                          marginRight: 5,
+                        }}>
+                        {truncateString(expense.description, 4)} ◦
+                      </PrimaryText>
+                    ) : null}
+
                     <PrimaryText
                       style={{
                         color: colors.primaryText,
@@ -207,14 +238,22 @@ const TransactionItem: React.FC<TransactionItemProps> = ({
                 </View>
               </View>
               <View>
-                <PrimaryText>
-                  {currencySymbol} {expense.amount}
+                <PrimaryText style={{fontSize: 13}}>
+                  {currencySymbol}
+                  {Number.isInteger(expense.amount)
+                    ? formatCurrency(expense.amount)
+                    : formatCurrency(expense.amount.toFixed(2))}
                 </PrimaryText>
               </View>
             </Animated.View>
           </Swipeable>
         </GestureHandlerRootView>
       ))}
+      <UndoModal
+        isVisible={deletedItem !== null}
+        onUndo={handleUndo}
+        type={'Transaction'}
+      />
     </View>
   );
 };
@@ -251,7 +290,7 @@ const TransactionList: React.FC<TransactionListProps> = ({
             nextWeek: 'dddd',
             lastDay: '[Yesterday]',
             lastWeek: '[Last] dddd',
-            sameElse: 'MMM YYYY',
+            sameElse: 'DD MMM YYYY',
           })}
         />
       ))}
